Remove protective clause for rsyslog_relays
In order to include all role defaults in the defaults file, an empty list
should be used for rsyslog_relays var instead of a commented-out example.

This removes superfluous protective clauses that obscure templates and
tasks, and ensures a user knows exactly which defaults are set when
investigating the role.